move camelizeData.ts into utils folder
